refactor(analytics): extract convertROIToCSV to sidecar (file-length governance)
Governance retro on the April 2026 8-PR merge batch flagged
ROIDashboardService.ts at 505 lines (over 500-line governance limit,
regression from SMI-1683 Wave 5B helper additions).

Fix: move the 39-line convertROIToCSV function into
ROIDashboardService.csv.ts. ROIDashboardService.ts now 467 lines with
headroom; the sidecar is 50 lines. Pure extraction — no behavior
change. Existing ROIDashboardService.test.ts 25/25 still pass.

Refs SMI-1683.
