+++ packages/core/src/analytics/ROIDashboardService.ts
@@ -11,6 +11,7 @@
 import type { Database as DatabaseType } from '../db/database-interface.js'
 import { ValidationError } from '../validation/validation-error.js'
 import { AnalyticsRepository } from './AnalyticsRepository.js'
+import { convertROIToCSV } from './ROIDashboardService.csv.js'
 import type { ROIDashboard, ROIMetrics, ExportFormat, UsageEvent } from './types.js'
 
 export interface ROIComputeOptions {
@@ -209,7 +210,7 @@ export class ROIDashboardService {
         return JSON.stringify(data, null, 2)
 
       case 'csv':
-        return this.convertROIToCSV(data)
+        return convertROIToCSV(data)
 
       case 'pdf':
         throw new Error('PDF export not yet implemented')
@@ -461,45 +462,5 @@ export class ROIDashboardService {
       estimatedValueUsd: 0,
       computedAt: new Date().toISOString(),
     }
-  }
-
-  private convertROIToCSV(data: Partial<ROIDashboard>): string {
-    const lines: string[] = []
-
-    if (data.user) {
-      lines.push('User ROI Dashboard')
-      lines.push('')
-      lines.push('Metric,Value')
-      lines.push(`User ID,${data.user.userId}`)
-      lines.push(`Total Time Saved (min),${data.user.totalTimeSaved.toFixed(1)}`)
-      lines.push(`Estimated Value (USD),${data.user.estimatedValueUsd.toFixed(2)}`)
-      lines.push('')
-      lines.push('Top Skills')
-      lines.push('Skill ID,Skill Name,Time Saved (min)')
-      for (const skill of data.user.topSkills) {
-        lines.push(`${skill.skillId},${skill.skillName},${skill.timeSaved.toFixed(1)}`)
-      }
-    }
-
-    if (data.stakeholder) {
-      lines.push('Stakeholder ROI Dashboard')
-      lines.push('')
-      lines.push('Metric,Value')
-      lines.push(`Total Users,${data.stakeholder.totalUsers}`)
-      lines.push(`Total Activations,${data.stakeholder.totalActivations}`)
-      lines.push(`Avg Time Saved Per User (min),${data.stakeholder.avgTimeSavedPerUser.toFixed(1)}`)
-      lines.push(`Total Estimated Value (USD),${data.stakeholder.totalEstimatedValue.toFixed(2)}`)
-      lines.push(`Adoption Rate (%),${(data.stakeholder.adoptionRate * 100).toFixed(1)}`)
-      lines.push('')
-      lines.push('Skill Leaderboard')
-      lines.push('Skill ID,Skill Name,User Count,Total Value (USD)')
-      for (const skill of data.stakeholder.skillLeaderboard) {
-        lines.push(
-          `${skill.skillId},${skill.skillName},${skill.userCount},${skill.totalValue.toFixed(2)}`
-        )
-      }
-    }
-
-    return lines.join('\n')
   }
 }
